v28.4.0 - merge from upstream 28.x and added several new features and Blackcoin specific adaptations. 

todo:
- create a good release notes
- translations 
- cleanup obsolete code
- adapt current tests
- add specific blackcoin tests
- switch to GUIX

+++ src/test/util/random.h
@@ -54,7 +54,7 @@ static inline bool InsecureRandBool()
 
 static inline CAmount InsecureRandMoneyAmount()
 {
-    return static_cast<CAmount>(InsecureRandRange(MAX_MONEY + 1));
+    return static_cast<CAmount>(InsecureRandRange(static_cast<uint64_t>(MAX_MONEY) + 1));
 }
 
 #endif // BITCOIN_TEST_UTIL_RANDOM_H
